# scada-reporter/frontend/openapi.json
{
  "components": {
    "schemas": {
...
      "Body_import_tags_api_tags_import_post": {
        "properties": {
          "file": {
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "type": "string"
          }
        },
        "required": [
          "file"
        ],
        "title": "Body_import_tags_api_tags_import_post",
        "type": "object"
      },
      "Body_import_tags_csv_api_tags_import_csv_post": {
        "properties": {
          "file": {
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "type": "string"
          }
        },
        "required": [
          "file"
        ],
        "title": "Body_import_tags_csv_api_tags_import_csv_post",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "TagCreate": {
        "properties": {
          "channel": {
            "default": "",
            "title": "Channel",
            "type": "string"
          },
          "data_type": {
            "default": "",
            "title": "Data Type",
            "type": "string"
          },
          "deadband": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deadband"
          },
          "description": {
            "default": "",
            "title": "Description",
            "type": "string"
          },
          "device": {
            "default": "",
            "title": "Device",
            "type": "string"
          },
          "long_term": {
            "default": false,
            "title": "Long Term",
            "type": "boolean"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "node_id": {
            "default": "",
            "title": "Node Id",
            "type": "string"
          },
          "plc_ip": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Plc Ip"
          },
          "plc_name": {
            "default": "",
            "title": "Plc Name",
            "type": "string"
          },
          "plc_rack": {
            "default": 0,
            "title": "Plc Rack",
            "type": "integer"
          },
          "plc_slot": {
            "default": 1,
            "title": "Plc Slot",
            "type": "integer"
          },
          "s7_address": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "S7 Address"
          },
          "sample_interval": {
            "default": 5,
            "title": "Sample Interval",
            "type": "integer"
          },
          "unit": {
            "default": "",
            "title": "Unit",
            "type": "string"
          }
        },
        "required": [
          "name"
        ],
        "title": "TagCreate",
        "type": "object"
      },
...
      "TagResponse": {
        "properties": {
          "channel": {
            "title": "Channel",
            "type": "string"
          },
          "current_value": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Value"
          },
          "daily_tracking": {
            "title": "Daily Tracking",
            "type": "boolean"
          },
          "data_type": {
            "title": "Data Type",
            "type": "string"
          },
          "deadband": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deadband"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "device": {
            "title": "Device",
            "type": "string"
          },
          "group_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Group Id"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "is_active": {
            "title": "Is Active",
            "type": "boolean"
          },
          "long_term": {
            "title": "Long Term",
            "type": "boolean"
          },
          "max_alarm": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Alarm"
          },
          "min_alarm": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Min Alarm"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "node_id": {
            "title": "Node Id",
            "type": "string"
          },
          "plc_ip": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Plc Ip"
          },
          "plc_name": {
            "title": "Plc Name",
            "type": "string"
          },
          "quality": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Quality"
          },
          "read_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Read At"
          },
          "s7_address": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "S7 Address"
          },
          "sample_interval": {
            "title": "Sample Interval",
            "type": "integer"
          },
          "unit": {
            "title": "Unit",
            "type": "string"
          }
        },
        "required": [
          "id",
          "node_id",
          "name",
          "description",
          "unit",
          "channel",
          "device",
          "plc_name",
          "plc_ip",
          "s7_address",
          "data_type",
          "sample_interval",
          "long_term",
          "daily_tracking",
          "is_active",
          "group_id",
          "min_alarm",
          "max_alarm",
          "deadband"
        ],
        "title": "TagResponse",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/auth/token"
          }
        },
        "type": "oauth2"
      }
    }
  },
  "info": {
    "description": "Su/Atıksu tesisi SCADA veri toplama ve raporlama sistemi",
    "title": "EKONT SMART REPORT",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/dashboard/tags": {
      "get": {
        "operationId": "dashboard_tags_api_dashboard_tags_get",
        "parameters": [
          {
            "in": "query",
            "name": "device",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Device"
            }
          },
          {
            "in": "query",
            "name": "search",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Search"
            }
          },
          {
            "in": "query",
            "name": "quality",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "good",
                    "bad",
                    "stale"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Quality"
            }
          },
          {
            "in": "query",
            "name": "daily",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Daily"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Page Size",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Dashboard Tags",
        "tags": [
          "dashboard"
        ]
      }
    },
...
    "/api/dashboard/watchlist-groups/{group_id}/tags/{tag_id}": {
      "delete": {
        "operationId": "remove_member_api_dashboard_watchlist_groups__group_id__tags__tag_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "group_id",
            "required": true,
            "schema": {
              "title": "Group Id",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "tag_id",
            "required": true,
            "schema": {
              "title": "Tag Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Remove Member",
        "tags": [
          "watchlist-groups"
        ]
      },
      "post": {
        "operationId": "add_member_api_dashboard_watchlist_groups__group_id__tags__tag_id__post",
        "parameters": [
          {
            "in": "path",
            "name": "group_id",
            "required": true,
            "schema": {
              "title": "Group Id",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "tag_id",
            "required": true,
            "schema": {
              "title": "Tag Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Add Member",
        "tags": [
          "watchlist-groups"
        ]
      }
    },
...
    "/api/tags/": {
      "get": {
        "operationId": "list_tags_api_tags__get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/TagResponse"
                  },
                  "title": "Response List Tags Api Tags  Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "List Tags",
        "tags": [
          "tags"
        ]
      },
      "post": {
        "operationId": "create_tag_api_tags__post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TagCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TagResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Create Tag",
        "tags": [
          "tags"
        ]
      }
    },
    "/api/tags/browse": {
      "get": {
        "description": "snap7 ile otomatik tag kesfi desteklenmez — bos liste doner.",
        "operationId": "browse_tags_api_tags_browse_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Browse Tags",
        "tags": [
          "tags"
        ]
      }
    },
    "/api/tags/export": {
      "get": {
        "description": "Tüm tag'leri CSV veya XLSX olarak indir. Aynı kolon düzeni import_csv ile uyumlu.",
        "operationId": "export_tags_api_tags_export_get",
        "parameters": [
          {
            "in": "query",
            "name": "format",
            "required": false,
            "schema": {
              "default": "csv",
              "pattern": "^(csv|xlsx)$",
              "title": "Format",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Export Tags",
        "tags": [
          "tags"
        ]
      }
    },
    "/api/tags/import": {
      "post": {
        "description": "WinCC full_export.xlsx yükle: Connection->IP çözülür, mutlak adres + tip ile\ntag'ler eklenir. Uzun-süre (archive) katalogu için `just seed-catalog` kullanın.",
        "operationId": "import_tags_api_tags_import_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_import_tags_api_tags_import_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Import Tags",
        "tags": [
          "tags"
        ]
      }
    },
    "/api/tags/import_csv": {
      "post": {
        "description": "Genel CSV import. En az `name` kolonu gerekir; node_id verilmezse türetilir.\nMevcut node_id atlanır. Kolonlar export ile aynıdır (eksikler varsayılan alır).",
        "operationId": "import_tags_csv_api_tags_import_csv_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_import_tags_csv_api_tags_import_csv_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Import Tags Csv",
        "tags": [
          "tags"
        ]
      }
    },
    "/api/tags/{tag_id}": {
      "delete": {
        "operationId": "delete_tag_api_tags__tag_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "tag_id",
            "required": true,
            "schema": {
              "title": "Tag Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Delete Tag",
        "tags": [
          "tags"
        ]
      },
      "patch": {
        "operationId": "update_tag_api_tags__tag_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "tag_id",
            "required": true,
            "schema": {
              "title": "Tag Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TagUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TagResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Update Tag",
        "tags": [
          "tags"
        ]
      }
    },
    "/api/tags/{tag_id}/readings": {
      "get": {
        "operationId": "get_readings_api_tags__tag_id__readings_get",
        "parameters": [
          {
            "in": "path",
            "name": "tag_id",
            "required": true,
            "schema": {
              "title": "Tag Id",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "start",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Start"
            }
          },
          {
            "in": "query",
            "name": "end",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "End"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 1000,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Get Readings",
        "tags": [
          "tags"
        ]
      }
    },
...
  }
}